@@ -1,60 +1,59 @@
 buildscript {
   dependencies {
-    classpath("net.serenity-bdd:serenity-gradle-plugin:2.4.34")
+    classpath "org.flywaydb:flyway-database-postgresql:11.0.1"
   }
 }
+
 plugins {
   id 'application'
-  id 'uk.gov.hmcts.java' version '0.12.61'
+  id 'uk.gov.hmcts.java' version '0.12.64'
   id 'pmd'
+  id 'idea'
   id 'jacoco'
-  id "org.flywaydb.flyway" version "7.15.0"
-  id 'io.spring.dependency-management' version '1.1.6'
-  id 'org.springframework.boot' version '3.3.4'
+  id 'project-report'
+  id "org.flywaydb.flyway" version '11.0.1'
+  id 'io.spring.dependency-management' version '1.1.7'
+  id 'org.springframework.boot' version '3.4.2'
   id 'com.github.ben-manes.versions' version '0.51.0'
-  id 'org.sonarqube' version '4.4.1.3373'
+  id 'org.sonarqube' version '6.0.1.5171'
   id "info.solidsoft.pitest" version '1.15.0'
-  id 'au.com.dius.pact' version '4.1.7' //do not change, otherwise serenity report fails
-  id 'org.owasp.dependencycheck' version '10.0.3'
+  id 'au.com.dius.pact' version '4.6.16'
+  id 'org.owasp.dependencycheck' version '11.1.1'
+  id 'net.serenity-bdd.serenity-gradle-plugin' version '4.2.16'
 }
-
-apply plugin: 'java'
-apply plugin: 'net.serenity-bdd.aggregator'
-apply plugin: 'project-report'
-apply plugin: 'idea'
 
 group = 'uk.gov.hmcts.reform'
 version = '0.0.1'
 
 java {
   toolchain {
-    languageVersion = JavaLanguageVersion.of(17)
+    languageVersion = JavaLanguageVersion.of(21)
   }
 }
 
 ext {
-  springCloudVersion = '2023.0.3'
+  springCloudVersion = '2024.0.0'
 }
 
 def versions = [
-  lombok          : '1.18.32',
-  gradlePitest    : '1.15.0',
-  pitest          : '1.16.3',
+  lombok          : '1.18.36',
   sonarPitest     : '0.5',
+  jjwt            : '0.12.6',
   junitPitest     : '1.2.1',
+  flywaydb        : '11.0.1',
   reformLogging   : '6.1.4',
-  pact_version    : '4.1.7',
-  log4j           : '2.23.1',
+  pact_version    : '4.6.16',
+  log4j           : '2.24.3',
   launchDarklySdk : "5.10.9",
-  reformS2sClient : '5.2.0',
-  springHystrix   : '2.2.10.RELEASE',
+  reformS2sClient : '5.3.0',
   restAssured     : '5.3.1',
   feign           : '3.8.0',
-  bytebuddy       : '1.14.18',
-  testContainer_postgresql: '1.20.1',
-  junit           : '5.9.2',
-  junitPlatform   : '1.10.3',
-  bouncycastle    : '1.78.1'
+  bytebuddy       : '1.15.10',
+  testContainer_postgresql: '1.20.4',
+  junit           : '5.11.3',
+  junitPlatform   : '1.11.3',
+  bouncycastle    : '1.79',
+  serenity        : '4.2.16'
 ]
 
 sourceSets {
