@@ -8,11 +8,11 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.boot.test.mock.mockito.SpyBean;
 import org.springframework.http.MediaType;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.context.bean.override.mockito.MockitoSpyBean;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import uk.gov.hmcts.reform.cdapi.domain.CaseFlag;
@@ -45,16 +45,16 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @WebMvcTest
 @WithMockUser
-@ContextConfiguration(classes = CaseFlagApiController.class)
+@ContextConfiguration(classes = { CaseFlagApiController.class, GlobalExceptionHandler.class})
 class CaseFlagApiControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
 
-    @SpyBean
+    @MockitoSpyBean
     private GlobalExceptionHandler globalExceptionHandler;
 
-    @MockBean
+    @MockitoBean
     private CaseFlagServiceImpl caseFlagService;
 
 
