@@ -5,11 +5,11 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.boot.test.mock.mockito.SpyBean;
 import org.springframework.http.MediaType;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.context.bean.override.mockito.MockitoSpyBean;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import uk.gov.hmcts.reform.cdapi.controllers.request.CategoryRequest;
@@ -37,13 +37,13 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @WebMvcTest
 @WithMockUser
-@ContextConfiguration(classes = CrdApiController.class)
+@ContextConfiguration(classes = { CrdApiController.class, GlobalExceptionHandler.class})
 class CrdApiControllerTest {
 
-    @MockBean
+    @MockitoBean
     CrdServiceImpl crdService;
 
-    @SpyBean
+    @MockitoSpyBean
     private GlobalExceptionHandler globalExceptionHandler;
 
     @Autowired
