@@ -77,7 +77,7 @@ class CaseFlagServiceImplTest {
     }
 
     @Test
-    void testGetCaseFlag_ByServiceIdAndWelshRequiredasY() {
+    void testGetCaseFlag_ByServiceIdAndWelshRequiredAsY() {
         when(caseFlagRepository.findAll(anyString())).thenReturn(getCaseFlagDtoList());
         assertThrows(ResourceNotFoundException.class, () ->
             caseFlagService.retrieveCaseFlagByServiceId("XXXX", "TEST", "y", "n"));
@@ -89,8 +89,7 @@ class CaseFlagServiceImplTest {
         "XXXX,PARTY,Y,true",
         "XXXX,PARTY,N,false"
     })
-    void testGetCaseFlags(String serviceId,String flagType,String welshRequired,
-                                                               boolean flag) {
+    void testGetCaseFlags(String serviceId, String flagType, String welshRequired, boolean flag) {
         when(caseFlagRepository.findAll(anyString()))
             .thenReturn(getCaseFlagDtoListWithLanguageInterpreter());
         when(listOfVenueRepository.findListOfValues(anyString()))
@@ -122,8 +121,9 @@ class CaseFlagServiceImplTest {
     void testGetCaseFlag_ByServiceIWithWelshRequired200(String serviceId, String welshRequired,
                                                                    String availableExternalFlag) {
         when(caseFlagRepository.findAll(anyString())).thenReturn(getCaseFlagDtoList());
-        var caseFlag = caseFlagService.retrieveCaseFlagByServiceId(serviceId, "", welshRequired,
-                                                                   availableExternalFlag);
+
+        var caseFlag = caseFlagService.retrieveCaseFlagByServiceId(
+            serviceId, "", welshRequired, availableExternalFlag);
         assertNotNull(caseFlag);
         assertEquals(1, caseFlag.getFlags().size());
         assertEquals(2, caseFlag.getFlags().get(0).getFlagDetails().size());
@@ -175,6 +175,19 @@ class CaseFlagServiceImplTest {
                 assertNotNull(flagDetail.getExternallyAvailable());
             }
         });
+    }
+
+    @Test
+    @DisplayName("Add child level flag adds nothing on special conditions")
+    void testAddChildLevelFlagContinuesWithoutAddingAnything() {
+        CaseFlagDto flagDto = new CaseFlagDto();
+        flagDto.setExternallyAvailable(false);
+        flagDto.setIsParent(false);
+        List<CaseFlagDto> dataList = List.of(flagDto);
+        List<FlagDetail> flagDetails = new ArrayList<>();
+
+        caseFlagService.addChildLevelFlag(dataList, flagDetails, "", true);
+        assertEquals(0, flagDetails.size());
     }
 
     List<CaseFlagDto> getEmptyCaseFlagDtoList(List<CaseFlagDto> caseFlagDtoList) {
@@ -313,78 +326,6 @@ class CaseFlagServiceImplTest {
 
         return caseFlagDtoList;
     }
-
-    private List<CaseFlagDto> getCaseFlagDtoListWithSignLanguage() {
-        var caseFlagDto1 = new CaseFlagDto();
-        caseFlagDto1.setFlagCode("CATEGORY");
-        caseFlagDto1.setCategoryId(0);
-        caseFlagDto1.setCategoryPath("");
-        caseFlagDto1.setId(1);
-        caseFlagDto1.setHearingRelevant(true);
-        caseFlagDto1.setRequestReason(false);
-        caseFlagDto1.setValueEn("PARTY");
-        caseFlagDto1.setValueCy("PARTY");
-        caseFlagDto1.setIsParent(true);
-        caseFlagDto1.setExternallyAvailable(false);
-        caseFlagDto1.setDefaultStatus("Requested");
-
-        var caseFlagDto2 = new CaseFlagDto();
-        caseFlagDto2.setFlagCode("RA0042");
-        caseFlagDto2.setCategoryId(1);
-        caseFlagDto2.setCategoryPath("PARTY");
-        caseFlagDto2.setValueCy("PARTY");
-        caseFlagDto2.setId(2);
-        caseFlagDto2.setHearingRelevant(true);
-        caseFlagDto2.setRequestReason(false);
-        caseFlagDto2.setValueEn("Sign Language");
-        caseFlagDto2.setValueCy("");
-        caseFlagDto2.setIsParent(false);
-        caseFlagDto1.setExternallyAvailable(true);
-        caseFlagDto1.setDefaultStatus("Approved");
-
-        var caseFlagDtoList = new ArrayList<CaseFlagDto>();
-        caseFlagDtoList.add(caseFlagDto1);
-        caseFlagDtoList.add(caseFlagDto2);
-
-        return caseFlagDtoList;
-    }
-
-    private List<CaseFlagDto> getCaseFlagDtoListWithOther() {
-        var caseFlagDto1 = new CaseFlagDto();
-        caseFlagDto1.setFlagCode("OT0001");
-        caseFlagDto1.setCategoryId(0);
-        caseFlagDto1.setCategoryPath("");
-        caseFlagDto1.setId(1);
-        caseFlagDto1.setHearingRelevant(true);
-        caseFlagDto1.setRequestReason(false);
-        caseFlagDto1.setValueEn("Other");
-        caseFlagDto1.setValueCy("Arall");
-        caseFlagDto1.setIsParent(true);
-        caseFlagDto1.setExternallyAvailable(false);
-        caseFlagDto1.setDefaultStatus("Requested");
-
-        var caseFlagDto2 = new CaseFlagDto();
-        caseFlagDto2.setFlagCode("RA0042");
-        caseFlagDto2.setCategoryId(1);
-        caseFlagDto2.setCategoryPath("PARTY");
-        caseFlagDto2.setValueCy("PARTY");
-        caseFlagDto2.setId(2);
-        caseFlagDto2.setHearingRelevant(true);
-        caseFlagDto2.setRequestReason(false);
-        caseFlagDto2.setValueEn("Sign Language");
-        caseFlagDto2.setValueCy("");
-        caseFlagDto2.setIsParent(false);
-        caseFlagDto1.setExternallyAvailable(true);
-        caseFlagDto1.setDefaultStatus("Approved");
-
-        var caseFlagDtoList = new ArrayList<CaseFlagDto>();
-        caseFlagDtoList.add(caseFlagDto1);
-        caseFlagDtoList.add(caseFlagDto2);
-
-        return caseFlagDtoList;
-    }
-
-
 
     private List<ListOfValue> getListOfValuesForLanguageInterPreter(boolean isWelshRequired) {
         var list = new ListOfValue();
